+++ lib/Data/Time/Format/Internal.hs
@@ -4,9 +4,10 @@
 --The contents of this module is liable to change, or disappear entirely.
 --Please <https://github.com/haskell/time/issues/new let me know> if you depend on anything here.
 module Data.Time.Format.Internal (
+    -- * ISO8601 formatting
     Format (..),
-    FormatTime (..),
-    ParseTime (..),
+    module Data.Time.Format.Format.Class,
+    module Data.Time.Format.Parse.Class,
 ) where
 
 import Data.Format